Adding recipe for CPAN hosted Perl5 module
Test::Directory at version 0.041.

Testing code can involve making sure that files are created and deleted as
expected. Doing this manually can be error prone, as it's easy to forget a
file, or miss that some unexpected file was added. This module simplifies
maintaining test directories by tracking their status as they are modified
or tested with this API, making it simple to test both individual files,
as well as to verify that there are no missing or unknown files.

Signed-off-by: Jens Rehsack <sno@netbsd.org>