Introduce recipe for CPAN hosted Perl module
constant::defer version 6 as prereq for Tim Orling's Slic3r goal.

constant::defer creates a subroutine which on the first call runs given
code to calculate its value, and on any subsequent calls just returns
that value, like a constant. The value code is discarded once run, allowing
it to be garbage collected.

Signed-off-by: Jens Rehsack <sno@netbsd.org>