feature-compat-try-perl: Add recipe for CPAN hosted Perl5 module
Add recipe for CPAN hosted Perl5 module Feature::Compat::Try version 0.04 as
prerequisite of upcoming update of Devel::MAT using Packager::Utils.

This module is written in preparation for when perl will gain true native
syntax support for try/catch control flow.  Perl added such syntax in the
development version 5.33.7, which is enabled by

    use feature 'try';

On that version of perl or later, this module simply enables the core
feature equivalent to using it directly. On such perls, this module will
install with no non-core dependencies, and requires no C compiler.

On older versions of perl before such syntax is available, it is currently
provided instead using the Syntax::Keyword::Try module, imported with a
special set of options to configure it to recognise exactly and only the
same syntax as the core perl feature, thus ensuring that any code using
it will still continue to function on that newer perl.

Signed-off-by: Jens Rehsack <sno@netbsd.org>