mozilla-publicsuffix-perl: Add recipe for CPAN hosted Perl5 module
Add recipe for CPAN hosted Perl5 module Mozilla::PublicSuffix version 1.0.6
as prerequisite of upcoming update of HTTP::CookieJar using Packager::Utils.

This module provides a single function that returns the public suffix of a
domain name by referencing a parsed copy of Mozilla's Public Suffix List.
From the official website at http://publicsuffix.org/:

    A "public suffix" is one under which Internet users can directly
    register names. Some examples of public suffixes are com, co.uk and
    pvt.k12.wy.us. The Public Suffix List is a list of all known public
    suffixes.

A copy of the official list is bundled with the distribution. As the
official list continues to be updated, the bundled copy will inevitably fall
out of date. Aside from new releases always including the latest version of
the list, this distribution's installer provides the option (defaults to
"No") to check for a new version of the list and download/use it if one is
found.

Signed-off-by: Jens Rehsack <sno@netbsd.org>