test-lib-perl: Add recipe for CPAN hosted Perl5 module
Add recipe for CPAN hosted Perl5 module Test::Lib version 0.002 as
prerequisite of upcoming import of CLI::Osprey using Packager::Utils.

Test::Lib searches upward from the calling module for a directory t with
a lib directory inside it, and adds it to the module search path. Looks
upward up to 5 directories. This is intended to be used in test modules
either directly in t or in a subdirectory to find their included testing
libraries located in t/lib.

Signed-off-by: Jens Rehsack <sno@netbsd.org>