digest-bubblebabble-perl: Add recipe for CPAN hosted Perl5 module
Add recipe for CPAN hosted Perl5 module Digest::BubbleBabble version 0.02 as
prerequisite of upcoming update of Net::DNS using Packager::Utils.

Digest::BubbleBabble takes a message digest (often generated by either of
the MD5 or SHA-1 message digest algorithms) and creates a fingerprint of
that digest in "bubble babble" format. Bubble babble is a method of
representing a message digest as a string of "real" words, to make the
fingerprint easier to remember. The "words" are not necessarily real
words, but they look more like words than a string of hex characters.

Signed-off-by: Jens Rehsack <sno@netbsd.org>